+++ app/src/main/java/org/shortweather/presentation/splash/SplashActivity.kt
@@ -20,7 +20,7 @@ import org.shortweather.util.extension.showToast
 @AndroidEntryPoint
 class SplashActivity : BindingActivity<ActivitySplashBinding>(R.layout.activity_splash) {
     private val viewModel by viewModels<InputTimeViewModel>()
-    private val SPLASH_TIME: Long = 3000  //3 sec
+    private val SPLASH_TIME: Long = 2800  //3 sec
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
